Follow-up commit for Homework 1 (26.09.2023)

+++ src/main/java/ru/skypro/homework/dto/Comment.java
@@ -10,7 +10,8 @@ public class Comment {
     private Integer author;
     private String authorImage;
     private String authorFirstName;
-    private Integer createdAt;
+    private Long createdAt;
+    //Исправил исходя из комментария на GitHub. Однако, в Schemas был Integer.
     private Integer pk;
     private String text;
